==> notebook/4_Data_Augmentation.ipynb <==
## 4 - preparation with augmentation

## classes with quite low mAP (classes to augment): 3, 23, 10, 30, 39, 19, 20, 21

### 3, 23 - not enough data - need to augment
### 19, 39 - confusion with 21 and 20 - perhaps bigger bounding box is needed - to be confirmed
### 10 - low amount of data, but mAP is good. 30 if labelled correctly should be close to it

!pip install scikit-learn

import os
import pandas as pd
import matplotlib.pyplot as plt
import matplotlib.image as mpimg
from sklearn.model_selection import train_test_split
import shutil
from PIL import Image, ImageDraw, ImageEnhance
import numpy as np
import random
# class, x_center, y_center, width, height

path = os.getcwd()

os.mkdir(path+'\\train')

os.mkdir(path+'\\val')

labels_df = pd.read_csv('df_keypoints_corrected.csv').drop('Unnamed: 0', axis = 1 )

labels_df

# Clipping and brightening

filenames_with_3 = []
for filename in labels_df['image_path'].unique():
    temp_df = labels_df[labels_df['image_path'] == filename]
    if (temp_df.loc[temp_df['kid'] == 3, 'vis'].values[0] == 2):
        filenames_with_3.append(filename)

        


len(filenames_with_3)

vis_val_df = pd.read_csv('vis_val_df_0.csv').drop('Unnamed: 0', axis = 1 )
vis_train_df = pd.read_csv('vis_train_df_0.csv').drop('Unnamed: 0', axis = 1 )

filenames_with_3_train = [item  for item in filenames_with_3 if item in vis_train_df['image_path'].unique()]
filenames_with_3_val = [item  for item in filenames_with_3 if item in vis_val_df['image_path'].unique()]

len(filenames_with_3_train)

len(filenames_with_3_val)

filenames_with_23 = []
for filename in labels_df['image_path'].unique():
    temp_df = labels_df[labels_df['image_path'] == filename]
    if (temp_df.loc[temp_df['kid'] == 23, 'vis'].values[0] == 2):
        filenames_with_23.append(filename)

        


len(filenames_with_23)

filenames_with_23_train = [item  for item in filenames_with_23 if item in vis_train_df['image_path'].unique()]
filenames_with_23_val = [item  for item in filenames_with_23 if item in vis_val_df['image_path'].unique()]

len(filenames_with_23_train)

len(filenames_with_23_val)

path3 = 'C:\\Users\\Kossa\\Desktop\\respo_vision\\yolov7-pitch\\respo_dataset\\data_pitch_geom\\recruitment_task_pitch_geom'

for img_name in filenames_with_3:

    img_name = img_name.split('/')[-1]

    print(img_name)

    img = mpimg.imread(path3+f'\\images\\train\\{img_name}')
    imgplot = plt.imshow(img)

    temp_df = labels_df[labels_df['image_path'] == f'images/train/{img_name}']
    
    # Annotation
    for row in temp_df.itertuples():
        if row.vis:
            if row.x > 1000:
                sign_x = -1
            else:
                sign_x = 1
            if row.y < 500:
                sign_y = 1
            else:
                sign_y = -1
                
            plt.annotate(row.kid, xy =(row.x , row.y ),
                            xytext =(row.x + sign_x * 100, row.y + sign_y * 100), 
                            color = 'red', 
                            arrowprops = dict(facecolor ='red',
                                              headwidth = 5,
                                              headlength = 5,
                                              width = 1,
                                              shrink = 0.00001),)
    
    plt.show()

# I do not want to overpopulate working classes, so I will select xClip (L and R) for each image, so it will only include points 3, 10 and 23, 30
# I will clip y as well
# skip 74f434bf605e8f6f5db0d409142bb7.jpg
xclip_3_10_R = [
    850,
    750,
    800,
    800,
    800,
    900,
    1100,
    800,
    750,
    850,
    900,
    900,
    900,
    850,
    850,
    900,
    900,
    1250,
    1100,
    750,
    750,
    800,
]

xclip_3_10_L = [
    50,
    0,
    10,
    50,
    20,
    30,
    100,
    80,
    0,
    10,
    0,
    150,
    140,
    30,
    60,
    10,
    40,
    70,
    50,
    20,
    30,
    120,
]

# yclip_3_10 random od 200 do 400 

xlip_for_filename_3_R = {filename:xclip for filename, xclip in zip(filenames_with_3, xclip_3_10_R)}

xlip_for_filename_3_L = {filename:xclip for filename, xclip in zip(filenames_with_3, xclip_3_10_L)}

xlip_for_filename_3_R

len(xclip_3_10_L)

for img_name in filenames_with_23:

    img_name = img_name.split('/')[-1]

    print(img_name)

    img = mpimg.imread(path3+f'\\images\\train\\{img_name}')
    imgplot = plt.imshow(img)

    temp_df = labels_df[labels_df['image_path'] == f'images/train/{img_name}']
    
    # Annotation
    for row in temp_df.itertuples():
        if row.vis:
            if row.x > 1000:
                sign_x = -1
            else:
                sign_x = 1
            if row.y < 500:
                sign_y = 1
            else:
                sign_y = -1
                
            plt.annotate(row.kid, xy =(row.x , row.y ),
                            xytext =(row.x + sign_x * 100, row.y + sign_y * 100), 
                            color = 'red', 
                            arrowprops = dict(facecolor ='red',
                                              headwidth = 5,
                                              headlength = 5,
                                              width = 1,
                                              shrink = 0.00001),)
    
    plt.show()

xclip_23_30 = [
    1100,
    750,
    1000,
    1100,
    900,
    1100,
    1200,
    1000,
    1100,
    900,
    900,
    1000,
    900,
    1150,
    750,
    1100,
    1150,
    1000,
    1000,
    1000,
    1000,
    1000,
    900,
    900,
    900,
    1300,
    1100,
    900,
    1100,
    1150,
    900,
    1100,
    1150,
    1150,
]
    
# y random 200 - 400 
    

xlip_for_filename_23 = {filename:xclip for filename, xclip in zip(filenames_with_23, xclip_23_30)}

xlip_for_filename_23

len(xclip_23_30)

## 2 ways to do this - clip and generate new - smaller image or gray everything else out (yolo will fit it to gray image anyway)

filenames_with_3

os.mkdir(path+'\\aug_3_10')

os.mkdir(path+'\\aug_3_10\\train')
os.mkdir(path+'\\aug_3_10\\val')

xlip_for_filename_3_R

filenames_with_3_train

img_name

random.seed(34)
random_clip_dict = {}

for img_name in filenames_with_3_train:

    brightness_start = 0.4

    for i in range(0,4):

        xclip_R = xlip_for_filename_3_R[img_name]
        xclip_L = xlip_for_filename_3_L[img_name]
        xclip_R = xclip_R + random.randint(-50,50)
        if xclip_L:
            xclip_L = xclip_L + random.randint(-50,50)
        yclip = random.randint(200,400)
    
        
        img_name_s = img_name.split('/')[-1]
    
        print(img_name)
    
        # img = mpimg.imread(path3+f'\\images\\train\\{img_name}')
        img = Image.open(path3+f'\\images\\train\\{img_name_s}')
        # imgplot = plt.imshow(img)
    
        temp_df = labels_df[labels_df['image_path'] == f'images/train/{img_name_s}']
        temp_df = temp_df[temp_df['x'] < xclip_R]
    
        # Size of the image in pixels (size of original image)
        # (This is not mandatory)
        width, height = img.size
         
        # Setting the points for cropped image
        left = xclip_L
        top = yclip
        right = xclip_R
        bottom = height
    
        
         
        # Cropped image of above dimension
        # (It will not change original image)
        im1 = img.crop((left, top, right, bottom))
    
        #image brightness enhancer
        enhancer = ImageEnhance.Brightness(im1)
        
        im2 = enhancer.enhance(brightness_start)
    
        filename = img_name_s[:-4]
    
        im2.save(f'aug_3_10\\train\\{filename}_{i}.jpg')
    
        random_clip_dict[f'aug_3_10\\train\\{filename}_{i}.jpg'] = [xclip_R, xclip_L, yclip]

        brightness_start += 0.4





random.seed(34)
# random_clip_dict = {}

for img_name in filenames_with_3_val:

    brightness_start = 0.4

    for i in range(0,4):

        xclip_R = xlip_for_filename_3_R[img_name]
        xclip_L = xlip_for_filename_3_L[img_name]
        xclip_R = xclip_R + random.randint(-50,50)
        if xclip_L:
            xclip_L = xclip_L + random.randint(-50,50)
        yclip = random.randint(200,400)
    
        
        img_name_s = img_name.split('/')[-1]
    
        print(img_name)
    
        # img = mpimg.imread(path3+f'\\images\\train\\{img_name}')
        img = Image.open(path3+f'\\images\\train\\{img_name_s}')
        # imgplot = plt.imshow(img)
    
        temp_df = labels_df[labels_df['image_path'] == f'images/train/{img_name_s}']
        temp_df = temp_df[temp_df['x'] < xclip_R]
    
        # Size of the image in pixels (size of original image)
        # (This is not mandatory)
        width, height = img.size
         
        # Setting the points for cropped image
        left = xclip_L
        top = yclip
        right = xclip_R
        bottom = height
    
        
         
        # Cropped image of above dimension
        # (It will not change original image)
        im1 = img.crop((left, top, right, bottom))
    
        #image brightness enhancer
        enhancer = ImageEnhance.Brightness(im1)
        
        im2 = enhancer.enhance(brightness_start)
    
        filename = img_name_s[:-4]
    
        im2.save(f'aug_3_10\\val\\{filename}_{i}.jpg')
    
        random_clip_dict[f'aug_3_10\\val\\{filename}_{i}.jpg'] = [xclip_R, xclip_L, yclip]

        brightness_start += 0.4





os.mkdir(path+'\\aug_23_30')

os.mkdir(path+'\\aug_23_30\\train')
os.mkdir(path+'\\aug_23_30\\val')

random.seed(34)
# random_clip_dict = {}

for img_name in filenames_with_23_train:

    brightness_start = 0.4

    for i in range(0,4):

        xclip = xlip_for_filename_23[img_name]
        xclip = xclip + random.randint(-50,50)
        yclip = random.randint(200,400)
    
        
        img_name_s = img_name.split('/')[-1]
    
        print(img_name)
    
        # img = mpimg.imread(path3+f'\\images\\train\\{img_name}')
        img = Image.open(path3+f'\\images\\train\\{img_name_s}')
        # imgplot = plt.imshow(img)
    
        temp_df = labels_df[labels_df['image_path'] == f'images/train/{img_name_s}']
        temp_df = temp_df[temp_df['x'] < xclip_R]
    
        # Size of the image in pixels (size of original image)
        # (This is not mandatory)
        width, height = img.size
         
        # Setting the points for cropped image
        left = xclip
        top = yclip
        right = 1920
        bottom = height
    
        
         
        # Cropped image of above dimension
        # (It will not change original image)
        im1 = img.crop((left, top, right, bottom))
    
        #image brightness enhancer
        enhancer = ImageEnhance.Brightness(im1)
        
        im2 = enhancer.enhance(brightness_start)
    
        filename = img_name_s[:-4]
    
        im2.save(f'aug_23_30\\train\\{filename}_{i}.jpg')
    
        random_clip_dict[f'aug_23_30\\train\\{filename}_{i}.jpg'] = [xclip,  yclip]

        brightness_start += 0.4





random.seed(34)
# random_clip_dict = {}

for img_name in filenames_with_23_val:

    brightness_start = 0.4

    for i in range(0,4):

        xclip = xlip_for_filename_23[img_name]
        xclip = xclip + random.randint(-50,50)
        yclip = random.randint(200,400)
    
        
        img_name_s = img_name.split('/')[-1]
    
        print(img_name)
    
        # img = mpimg.imread(path3+f'\\images\\train\\{img_name}')
        img = Image.open(path3+f'\\images\\train\\{img_name_s}')
        # imgplot = plt.imshow(img)
    
        temp_df = labels_df[labels_df['image_path'] == f'images/train/{img_name_s}']
        temp_df = temp_df[temp_df['x'] < xclip_R]
    
        # Size of the image in pixels (size of original image)
        # (This is not mandatory)
        width, height = img.size
         
        # Setting the points for cropped image
        left = xclip
        top = yclip
        right = 1920
        bottom = height
    
        
         
        # Cropped image of above dimension
        # (It will not change original image)
        im1 = img.crop((left, top, right, bottom))
    
        #image brightness enhancer
        enhancer = ImageEnhance.Brightness(im1)
        
        im2 = enhancer.enhance(brightness_start)
    
        filename = img_name_s[:-4]
    
        im2.save(f'aug_23_30\\val\\{filename}_{i}.jpg')
    
        random_clip_dict[f'aug_23_30\\val\\{filename}_{i}.jpg'] = [xclip,  yclip]

        brightness_start += 0.4





## generating labels: fall lables should be recalculated

os.mkdir(path+'\\aug_3_10\\train\\labels')
os.mkdir(path+'\\aug_3_10\\val\\labels')
os.mkdir(path+'\\aug_23_30\\train\\labels')
os.mkdir(path+'\\aug_23_30\\val\\labels')

path

# <object-class-id> <x> <y> <width> <height>
label_pad = 50 #px
# label_pad_30 = 80 # in fact I will increase bounding box for class 30
for img_name in filenames_with_3_train:
# for df in vis_train_df.groupby('image_path'):
    # print(df[1]['x'])
    # xclip = xlip_for_filename_3[img_name]

    filename = img_name.split('/')[-1][:-4]

    # print(filename, '\n')
    for i in range(4):
        xclip_R, xclip_L, yclip = random_clip_dict[f'aug_3_10\\train\\{filename}_{i}.jpg']
        temp_df = vis_train_df[(vis_train_df['image_path'] == img_name) & (vis_train_df['x'] < xclip_R)]
        
        with open(f'{path}\\aug_3_10\\train\\labels\\{filename}_{i}.txt', 'w') as file:
            for row in temp_df.itertuples():
                # x_l = row.x - label_pad
                # x_p = row.x + label_pad
                # if x_l < 0: x_l = 0
                # if x_p > 1920: x_l = 1920
                # y_t = row.y - label_pad
                # y_b = row.y + label_pad
                # if y_t < 0: y_t = 0
                # if y_b > 1080: y_b = 1080
                if not row.vis:
                    continue
                else:
                    file.write(f'{row.kid} {(row.x - xclip_L)/(xclip_R - xclip_L)} {(row.y - yclip)/(1080 - yclip)} {2*label_pad/(xclip_R - xclip_L)} {2*label_pad/(1080 - yclip)}\n')
                # print((f'{row.kid} {row.x/1920} {row.y/1080} {2*label_pad/1920} {2*label_pad/1080}'))
            
        # print(rx_l = row.x - 50)
        # print(row.y)
    # print(object_class_id)

random_clip_dict

# <object-class-id> <x> <y> <width> <height>
label_pad = 50 #px
# label_pad_30 = 80 # in fact I will increase bounding box for class 30
for img_name in filenames_with_3_val:
# for df in vis_train_df.groupby('image_path'):
    # print(df[1]['x'])
    
    filename = img_name.split('/')[-1][:-4]

    # print(filename, '\n')
    for i in range(4):
        xclip_R, xclip_L, yclip = random_clip_dict[f'aug_3_10\\val\\{filename}_{i}.jpg']
        temp_df = vis_val_df[(vis_val_df['image_path'] == img_name) & (vis_val_df['x'] < xclip_R)]
        
        with open(f'{path}\\aug_3_10\\val\\labels\\{filename}_{i}.txt', 'w') as file:
            for row in temp_df.itertuples():
                # x_l = row.x - label_pad
                # x_p = row.x + label_pad
                # if x_l < 0: x_l = 0
                # if x_p > 1920: x_l = 1920
                # y_t = row.y - label_pad
                # y_b = row.y + label_pad
                # if y_t < 0: y_t = 0
                # if y_b > 1080: y_b = 1080
                if not row.vis:
                    continue
                else:
                    file.write(f'{row.kid} {(row.x - xclip_L)/(xclip_R - xclip_L)} {(row.y - yclip)/(1080 - yclip)} {2*label_pad/(xclip_R - xclip_L)} {2*label_pad/(1080 - yclip)}\n')
                # print((f'{row.kid} {row.x/1920} {row.y/1080} {2*label_pad/1920} {2*label_pad/1080}'))
            
        # print(rx_l = row.x - 50)
        # print(row.y)
    # print(object_class_id)

# random.seed(0)

class_name_to_id_mapping = {str(i):i for i in range(40)}

class_id_to_name_mapping = dict(zip(class_name_to_id_mapping.values(), class_name_to_id_mapping.keys()))

def plot_bounding_box(image, annotation_list):
    annotations = np.array(annotation_list)
    w, h = image.size
    
    plotted_image = ImageDraw.Draw(image)

    transformed_annotations = np.copy(annotations)
    transformed_annotations[:,[1,3]] = annotations[:,[1,3]] * w
    transformed_annotations[:,[2,4]] = annotations[:,[2,4]] * h 
    
    transformed_annotations[:,1] = transformed_annotations[:,1] - (transformed_annotations[:,3] / 2)
    transformed_annotations[:,2] = transformed_annotations[:,2] - (transformed_annotations[:,4] / 2)
    transformed_annotations[:,3] = transformed_annotations[:,1] + transformed_annotations[:,3]
    transformed_annotations[:,4] = transformed_annotations[:,2] + transformed_annotations[:,4]
    
    for ann in transformed_annotations:
        obj_cls, x0, y0, x1, y1 = ann
        plotted_image.rectangle(((x0,y0), (x1,y1)))
        
        plotted_image.text((x0, y0 - 10), class_id_to_name_mapping[(int(obj_cls))])
    
    plt.imshow(np.array(image))
    plt.show()

for filename in filenames_with_3_train:
    
    
    filename = filename.split('/')[-1][:-4]

    for i in range(4):

        img_path = f'{path}\\aug_3_10\\train\\{filename}_{i}.jpg'
        labels_path = f'{path}\\aug_3_10\\train\\labels\\{filename}_{i}.txt'
    
            
        
        image = Image.open(img_path)
        with open(labels_path, 'r') as file:
            labels = file.readlines()
    
        
        annotation_file = labels_path
        with open(annotation_file, "r") as file:
            annotation_list = file.read().split("\n")[:-1]
            annotation_list = [x.split(" ") for x in annotation_list]
            annotation_list = [[float(y) for y in x ] for x in annotation_list]
    
        plot_bounding_box(image, annotation_list)

for filename in filenames_with_3_val:
    
    filename = filename.split('/')[-1][:-4]

    for i in range(4):
    
        img_path = f'{path}\\aug_3_10\\val\\{filename}_{i}.jpg'
        labels_path = f'{path}\\aug_3_10\\val\\labels\\{filename}_{i}.txt'
        
        image = Image.open(img_path)
        with open(labels_path, 'r') as file:
            labels = file.readlines()
        
        annotation_file = labels_path
        with open(annotation_file, "r") as file:
            annotation_list = file.read().split("\n")[:-1]
            annotation_list = [x.split(" ") for x in annotation_list]
            annotation_list = [[float(y) for y in x ] for x in annotation_list]
            
        # arr_labels = np.array([b_list]).reshape(len(b_list), 5)
        #Plot the Bounding Box
        plot_bounding_box(image, annotation_list)

# <object-class-id> <x> <y> <width> <height>
label_pad = 50 #px
# label_pad_30 = 80 # in fact I will increase bounding box for class 30
for img_name in filenames_with_23_train:

    filename = img_name.split('/')[-1][:-4]

    # print(filename, '\n')
    for i in range(4):
        xclip_L, yclip = random_clip_dict[f'aug_23_30\\train\\{filename}_{i}.jpg']
        temp_df = vis_train_df[(vis_train_df['image_path'] == img_name) & (vis_train_df['x'] > xclip_L)]
        
        with open(f'{path}\\aug_23_30\\train\\labels\\{filename}_{i}.txt', 'w') as file:
            for row in temp_df.itertuples():

                if not row.vis:
                    continue
                else:
                    file.write(f'{row.kid} {(row.x - xclip_L)/(1920 - xclip_L)} {(row.y - yclip)/(1080 - yclip)} {2*label_pad/(1920 - xclip_L)} {2*label_pad/(1080 - yclip)}\n')
                # print((f'{row.kid} {row.x/1920} {row.y/1080} {2*label_pad/1920} {2*label_pad/1080}'))


# <object-class-id> <x> <y> <width> <height>
label_pad = 50 #px
# label_pad_30 = 80 # in fact I will increase bounding box for class 30
for img_name in filenames_with_23_val:

    filename = img_name.split('/')[-1][:-4]

    for i in range(4):
        xclip_L, yclip = random_clip_dict[f'aug_23_30\\val\\{filename}_{i}.jpg']
        temp_df = vis_val_df[(vis_val_df['image_path'] == img_name) & (vis_val_df['x'] > xclip_L)]
        
        with open(f'{path}\\aug_23_30\\val\\labels\\{filename}_{i}.txt', 'w') as file:
            for row in temp_df.itertuples():

                if not row.vis:
                    continue
                else:
                    file.write(f'{row.kid} {(row.x - xclip_L)/(1920 - xclip_L)} {(row.y - yclip)/(1080 - yclip)} {2*label_pad/(1920 - xclip_L)} {2*label_pad/(1080 - yclip)}\n')
                # print((f'{row.kid} {row.x/1920} {row.y/1080} {2*label_pad/1920} {2*label_pad/1080}'))


for filename in filenames_with_23_train:
    
    filename = filename.split('/')[-1][:-4]

    for i in range(4):

        img_path = f'{path}\\aug_23_30\\train\\{filename}_{i}.jpg'
        labels_path = f'{path}\\aug_23_30\\train\\labels\\{filename}_{i}.txt'
    
        image = Image.open(img_path)
        with open(labels_path, 'r') as file:
            labels = file.readlines()
    
        
        annotation_file = labels_path
        with open(annotation_file, "r") as file:
            annotation_list = file.read().split("\n")[:-1]
            annotation_list = [x.split(" ") for x in annotation_list]
            annotation_list = [[float(y) for y in x ] for x in annotation_list]
            
        # arr_labels = np.array([b_list]).reshape(len(b_list), 5)
        #Plot the Bounding Box
        plot_bounding_box(image, annotation_list)

for filename in filenames_with_23_val:
    
    filename = filename.split('/')[-1][:-4]

    for i in range(4):

        img_path = f'{path}\\aug_23_30\\val\\{filename}_{i}.jpg'
        labels_path = f'{path}\\aug_23_30\\val\\labels\\{filename}_{i}.txt'
    
        image = Image.open(img_path)
        with open(labels_path, 'r') as file:
            labels = file.readlines()
    
        
        annotation_file = labels_path
        with open(annotation_file, "r") as file:
            annotation_list = file.read().split("\n")[:-1]
            annotation_list = [x.split(" ") for x in annotation_list]
            annotation_list = [[float(y) for y in x ] for x in annotation_list]
            
        # arr_labels = np.array([b_list]).reshape(len(b_list), 5)
        #Plot the Bounding Box
        plot_bounding_box(image, annotation_list)

## first stage - brightening and multiplaying images for critical classes is ready - images will be added to training and validating datasets

## Second idea - perhaps bigger bounding boxes are needed for problematic classes

labels_df = labels_df[labels_df['dataset'] == 'train']

labels_df

all_images = labels_df['image_path'].unique()

# all_images = [item.split('/')[-1] for item in all_images]

train, val = train_test_split(all_images, test_size=0.2, random_state=43, shuffle=True)


len(train)

len(val)

train_df = labels_df[labels_df['image_path'].isin(train)]

val_df = labels_df[labels_df['image_path'].isin(val)]

val_df

val_df.to_csv('val_df_0.csv')

vis_train_df = train_df[train_df['vis']==2]
vis_val_df = val_df[val_df['vis']==2]

vis_val_df.to_csv('vis_val_df_0.csv')

len(vis_val_df) / len(vis_train_df) # in fact it is 0.244 split but its ok

plt.figure(figsize = (16, 4))
plt.bar(vis_train_df.groupby('kid').count()['x'].index, vis_train_df.groupby('kid').count()['x'])
plt.xticks(rotation = 45)
plt.xticks(vis_train_df.groupby('kid').count()['x'].index)
plt.show()

plt.figure(figsize = (16, 4))
plt.bar(vis_val_df.groupby('kid').count()['x'].index, vis_val_df.groupby('kid').count()['x'])
plt.xticks(rotation = 45)
plt.xticks(vis_val_df.groupby('kid').count()['x'].index)
plt.show()

## class balance comparison between train and test is fine

#check for leakage
for item in train:
    if item in val:
        print("!")
for item in val:
    if item in train:
        print("!")

os.mkdir(path+'\\train\\images')
os.mkdir(path+'\\train\\labels')
os.mkdir(path+'\\val\\images')
os.mkdir(path+'\\val\\labels')

path2 = 'C:\\Users\\Kossa\\Desktop\\respo_vision\\yolov7-pitch\\respo_dataset\\data_pitch_geom\\recruitment_task_pitch_geom'

# for img_path in train:
#     filename = img_path.split('/')[-1]
#     shutil.copy(path2+f'\\images\\train\\{filename}', path+f'\\train\\images\\{filename}')


# Some files are mislabeled as training data, but they do not have any labels, despite the points are visible. 
# I will ignore them and check for vis == 2

for img_path in vis_train_df['image_path'].unique():
    filename = img_path.split('/')[-1]
    shutil.copy(path2+f'\\images\\train\\{filename}', path+f'\\train\\images\\{filename}')

# for img_path in val:
#     filename = img_path.split('/')[-1]
#     shutil.copy(path2+f'\\images\\train\\{filename}', path+f'\\val\\images\\{filename}')

for img_path in vis_val_df['image_path'].unique():
    filename = img_path.split('/')[-1]
    shutil.copy(path2+f'\\images\\train\\{filename}', path+f'\\val\\images\\{filename}')


no_vis_classes_train = []
for img_path in train:
    filename = img_path.split('/')[-1]
    # shutil.copy(path2+f'\\images\\train\\{filename}', path+f'\\val\\images\\{filename}')
    if filename not in [item.split('/')[-1] for item in vis_train_df['image_path'].unique()]:
        no_vis_classes_train.append(filename)

no_vis_classes_train


no_vis_classes_val = []
for img_path in val:
    filename = img_path.split('/')[-1]
    # shutil.copy(path2+f'\\images\\train\\{filename}', path+f'\\val\\images\\{filename}')
    if filename not in [item.split('/')[-1] for item in vis_val_df['image_path'].unique()]:
        no_vis_classes_val.append(filename)

no_vis_classes_val

# this includes above no vis classes - useful for testing but not for training
for img_path in val:
    filename = img_path.split('/')[-1]
    shutil.copy(path2+f'\\images\\train\\{filename}', path+f'\\val\\images\\{filename}')

# <object-class-id> <x> <y> <width> <height>
cls_with_bigger_box = [3, 23, 10, 30, 39, 19, 20, 21]
label_pad = 50 #px
bigger_label_pad = 80 # 60 % bigger, becasue its x2
for df in vis_train_df.groupby('image_path'):
    # print(df[1]['x'])
    filename = df[0].split('/')[-1][:-4]
    # print(filename, '\n')
    with open(f'{path}\\train\\labels\\{filename}.txt', 'w') as file:
        for row in df[1].itertuples():

            if not row.vis:
                continue
            else:
                if row.kid in cls_with_bigger_box:
                    file.write(f'{row.kid} {row.x/1920} {row.y/1080} {2*bigger_label_pad/1920} {2*bigger_label_pad/1080}\n')
                else:
                    file.write(f'{row.kid} {row.x/1920} {row.y/1080} {2*label_pad/1920} {2*label_pad/1080}\n')


# <object-class-id> <x> <y> <width> <height>
label_pad = 50 #px
bigger_label_pad = 80
# for df in vis_val_df.groupby('image_path'):    # for training
for df in val_df.groupby('image_path'):    # for testing
    # print(df[1]['x'])
    filename = df[0].split('/')[-1][:-4]
    with open(f'{path}\\val\\labels\\{filename}.txt', 'w') as file:
        for row in df[1].itertuples():

            if not row.vis:
                continue
            else:
                if row.kid in cls_with_bigger_box:
                    file.write(f'{row.kid} {row.x/1920} {row.y/1080} {2*bigger_label_pad/1920} {2*bigger_label_pad/1080}\n')
                else:
                    file.write(f'{row.kid} {row.x/1920} {row.y/1080} {2*label_pad/1920} {2*label_pad/1080}\n')
        
        # print(rx_l = row.x - 50)
        # print(row.y)
    # print(object_class_id)

# random.seed(0)

class_name_to_id_mapping = {str(i):i for i in range(40)}

class_id_to_name_mapping = dict(zip(class_name_to_id_mapping.values(), class_name_to_id_mapping.keys()))

def plot_bounding_box(image, annotation_list):
    annotations = np.array(annotation_list)
    w, h = image.size
    
    plotted_image = ImageDraw.Draw(image)

    transformed_annotations = np.copy(annotations)
    transformed_annotations[:,[1,3]] = annotations[:,[1,3]] * w
    transformed_annotations[:,[2,4]] = annotations[:,[2,4]] * h 
    
    transformed_annotations[:,1] = transformed_annotations[:,1] - (transformed_annotations[:,3] / 2)
    transformed_annotations[:,2] = transformed_annotations[:,2] - (transformed_annotations[:,4] / 2)
    transformed_annotations[:,3] = transformed_annotations[:,1] + transformed_annotations[:,3]
    transformed_annotations[:,4] = transformed_annotations[:,2] + transformed_annotations[:,4]
    
    for ann in transformed_annotations:
        obj_cls, x0, y0, x1, y1 = ann
        plotted_image.rectangle(((x0,y0), (x1,y1)))
        
        plotted_image.text((x0, y0 - 10), class_id_to_name_mapping[(int(obj_cls))])
    
    plt.imshow(np.array(image))
    plt.show()

filename = '0a996c430c38f452eb39b86b0b9b07'
img_path = f'{path}\\train\\images\\{filename}.jpg'
labels_path = f'{path}\\train\\labels\\{filename}.txt'

image = Image.open(img_path)
with open(labels_path, 'r') as file:
    labels = file.readlines()

annotation_file = labels_path
with open(annotation_file, "r") as file:
    annotation_list = file.read().split("\n")[:-1]
    annotation_list = [x.split(" ") for x in annotation_list]
    annotation_list = [[float(y) for y in x ] for x in annotation_list]
    
# arr_labels = np.array([b_list]).reshape(len(b_list), 5)
#Plot the Bounding Box
plot_bounding_box(image, annotation_list)

filename = '16bafd79da7d6a6c2e6893f1ffb4d9'
img_path = f'{path}\\train\\images\\{filename}.jpg'
labels_path = f'{path}\\train\\labels\\{filename}.txt'

image = Image.open(img_path)
with open(labels_path, 'r') as file:
    labels = file.readlines()

annotation_file = labels_path
with open(annotation_file, "r") as file:
    annotation_list = file.read().split("\n")[:-1]
    annotation_list = [x.split(" ") for x in annotation_list]
    annotation_list = [[float(y) for y in x ] for x in annotation_list]
    
# arr_labels = np.array([b_list]).reshape(len(b_list), 5)
#Plot the Bounding Box
plot_bounding_box(image, annotation_list)

## last step is to check how bounding boxes fit estimated point 30 position

fix_30_list = [
'9677b5248c7d2a391bb331263538c8.jpg',
'4fca4e0d72eef70eb308f4e92931b3.jpg',
'74f434bf605e8f6f5db0d409142bb7.jpg',
'fa9255414e8e18aa257670935fffba.jpg',
'60d86f695f30c394960023f391f50d.jpg',
'b73ab787697aec7a29675a88e771da.jpg',
'5fcf8d794a1c34024cc2c9efc285d9.jpg',
'b5d9188836d57ad32c0820713ab33e.jpg',
'41e9761846a0f10a8e6aecef400413.jpg',
'8dea69d0d05031d04d2b85af2a0705.jpg',
'2f1d6e937128e45c8257bf4b76db2e.jpg',
'cbf9e8794cd6aa406d2fc18946b89c.jpg',
'c05f429dbf4abeebeb4fa2e4937db8.jpg',
'2abe6d162668bb85bd8844103d58b3.jpg',
'ab65a0ffacb84533cf601a236f0920.jpg',
'38829ff54c7b1db3f49a31a1d5ece5.jpg',
'236e289ccb95432f1bbe07cf106c3b.jpg',
'2ef80b5e833ab9a2a3e269489dbf8d.jpg',
'a5941c096e19a477c13bb973e09df7.jpg',
'63cb5a548cd1032174d6605fabd290.jpg',
'4ef2046190bda2894769bf65e33ce7.jpg',
'3a4218907b274807c74be14032c901.jpg',
'78605986e67c9b89f1a807e858fe29.jpg',
'0676caa13e8abe3a568a2c61ef6e2d.jpg',
'f658ca5bdf4d8ee04fa3d648c9cc4f.jpg',
'7b304f4cd963b2fe573c3840315e79.jpg',
'66fa19216e23d7596b4d90da268e7c.jpg',# 33 confusion!!!!!!
'a1401f391b8f64656723df05c36b24.jpg',
'2791f191af7fca4b2e0b77cd050629.jpg',
'86a2a40a485feed061b6c5b5f961ee.jpg',
'c4f1f57482859c60ef0742691607f9.jpg',
'9ed1c1c9ee6cb5d5dbd61727d1c412.jpg',
'67d48e85a7792f3823b38c20a8ebf3.jpg',
'29c16eecccde846baba0b4f8bb5027.jpg',
'6546c56f291892f9b07a83eb2b1994.jpg',
'fdbab61c397896251067d8becc67e6.jpg',
'56c34b486748121bc22586949b0692.jpg',
'cdf10c1766c99298fed520400472a0.jpg',
'05a0d9cef3fb5fca8757e07b3bd867.jpg',
'a7267b1486cf8f8f8705e32452199e.jpg',
'ea158fb0fe07da57f02ff65684a647.jpg',
'335155589d1642be50e4de8953a6bf.jpg',
'2356ce9513bb1fe22364646639add7.jpg',
'9f8359d4bc8d2e63a0ce666f1936fd.jpg',
'9997dcea7cf66fe6e57fb36233208d.jpg',
'472fded23adc33c9a8bd5f7cbfd6b8.jpg',
'48d469d5006ce73f21e4be55a4f706.jpg',
'd4e0605f208a8365540f188d419ca6.jpg',
'e575b9605fb6f9fd6f729f393398b3.jpg',
'e25802bee03661511ddf0813904fea.jpg',
'a270c12019e0005c93e28ca45b06dd.jpg',
'a6eddc8c6d6d2afdbc871bf39ad49d.jpg',
'48f9c434173add06c222a44b3a95e4.jpg',
'62de8876e9adc7aa23af2ab13ec754.jpg',
'5f9a45153673ad24eb9f4ce6654ff1.jpg',
'0d0365deb03c4703e978f360d5ff72.jpg',
'710728f1a99225235e22abc62aeaab.jpg',
'02015eb88e33614b36170db06784d5.jpg',
'164bc44b724a585ae72e69ace0a10b.jpg',
'385acceeae6be647e931135c06bbea.jpg',
'2c4ff6598a91a53d7706ac185df094.jpg',
'92ffc7f0174f2e1e88ea739aef1f3f.jpg',
'a2cd5ab3d4985525865e565111eeeb.jpg',
'9bda0e6d253f33d80d4611f795a0c9.jpg',
'8f1bba299aa3484fc57a5a685e7be1.jpg',
'54fbd9767485472fa92a665b5ad9ff.jpg',
'2faca0c080867d91e5f7991b19f845.jpg',
'172dca5da5144cdb86558b9a6f1308.jpg',
'6e1f06d09ea276cc6c2e0d233facc3.jpg',
'd1f64e3430882e41de16b6b879fd9b.jpg',
'09f0bf76cd672ba841c926d8fdbfc8.jpg',
'03cc1aaab2c0985634cf4fa638b63a.jpg',
'cee6864c0ab20423896373ee8328d3.jpg',
'245f3357b3f3c7e97278610664ab8b.jpg',  ### specific orientation
'e681aa8ce2ce22486fd5d6d5870d2d.jpg',
'aef893839b0d08bc307f6c5b581c95.jpg',
'db6d9c9a386a306a65ed3939b65816.jpg',
'33b745960e851d2fd95e4abd6be141.jpg',
'0333a911c259ffb96e045cc998657f.jpg',
'75082ed4b6d5cf01ca376bbf67bcd8.jpg',
'00e6fafafd740664bef9aa4f387359.jpg',
'2c90e1733e23e5e56dbd63db45c63d.jpg',
'24e341f3e6399b9ad5864374a2a3c9.jpg',
'879b988c4b4d8f07709e2fd60510cf.jpg',
'05774d01da335be8ca3e263d622fdb.jpg',
'b5c506b0396c461aacf3c97e64afcf.jpg',
'a64c0ae39bb932c6a2eee247170c61.jpg',
'57de493eb70d824253227ddddbad86.jpg',
'17c0fd4742d6f872316ee1dd4959b8.jpg',
'21a05750e2d5fda4e77ebb179b86ea.jpg',
'1d2f4e3883651bb8a261cd1df3fcb6.jpg', ### specific orientation
'94b49ca68ed419da8ddbccd73dfd03.jpg', ### specific orientation
'fc8446a9d0ccde9c5d6cb10c6d3883.jpg',
'679fe17d8e3159e151938b5d8b3d83.jpg',
'b31a75c2688625e447879c620ac151.jpg',
'98b8d1e43aa64028280da5c354d459.jpg',
'65342549664333e6167919ff91fe47.jpg',
'06852b013d652b6850cba20fff83d8.jpg',
'bfc68987f3a82449d5533cb046deef.jpg',
'6ac91cbebe6df9a2669b1e0d008bbf.jpg',
'41375ecb2f0390f4328cb1c9a94179.jpg',
'baf7ca6c847404d9962ed81260c5c7.jpg',
'a1f6ede2ad8d82eeaf59b2dc022da8.jpg',
'28576c4dc0ff8c7082b63e6da72442.jpg',
'96c179f7e57d9e6204bde75aac9778.jpg',
'316777dee32535be262f63de39ca2d.jpg',
'b8f00314e9c80fc3d2a609adb41e63.jpg',
'f347fafdc58e6081b56ab63008dcfc.jpg',
'2869f77b076470b28b5f1b26b9a8df.jpg',
'665eed70d66d36754e1f4ff3688e6c.jpg',
'f1f9b766b7eff4d6de3792a89184fd.jpg',
'c91f4f418ead50690954cf487bf877.jpg',
'b2824f7190be843397ef4e2cd5cbdf.jpg',
'1e21d02ee9d25d2318964fd4bae782.jpg',
'c9819a80d876f2b4f9464a067eb65a.jpg',
'fc673285d41e0ef79c9bd8c887565c.jpg',
'32fcc238f33bb7165f3fa86c17620d.jpg',
'76b8e63b273ab668dfcef2805d92c5.jpg',
'10642681aebe3a3312d4fe3f6f4f1f.jpg',
]

train

for filename in fix_30_list:
    if f'images/train/{filename}' in train:
        filename = filename.split('.')[0]
        print(filename)
        img_path = f'{path}\\train\\images\\{filename}.jpg'
        labels_path = f'{path}\\train\\labels\\{filename}.txt'
    else:
        filename = filename.split('.')[0]
        print(filename)
        img_path = f'{path}\\val\\images\\{filename}.jpg'
        labels_path = f'{path}\\val\\labels\\{filename}.txt'
        
    
    image = Image.open(img_path)
    with open(labels_path, 'r') as file:
        labels = file.readlines()
    
    
    annotation_file = labels_path
    with open(annotation_file, "r") as file:
        annotation_list = file.read().split("\n")[:-1]
        annotation_list = [x.split(" ") for x in annotation_list]
        annotation_list = [[float(y) for y in x ] for x in annotation_list]
        
    # arr_labels = np.array([b_list]).reshape(len(b_list), 5)
    #Plot the Bounding Box
    plot_bounding_box(image, annotation_list)

# it is decent, but some labels could be corrected in the future

filenames_with_3_val

filenames_with_23_val
